Más fixes
- Error al volver al menú principal solucionado
- El menú principal funciona correctamente cuando vuelves a él, si no
hay publicaciones nuevas, lo muestra
- El icono de los álbumes se genera al crear el álbum, lo que hace la
navegación entre álbumes mucho más rápida
- Arreglado el icono de los álbumes, no se mostraba correctamente por
las dimensiones

+++ PhotoTDS/src/umu/tds/view/AddPublicationWindow.java
@@ -328,10 +328,12 @@ public class AddPublicationWindow implements IEncendidoListener {
 	    String nombreArchivo = rutaOrigen.getFileName().toString();
 	    Path rutaDestino = Paths.get("src/umu/tds/photos/" + nombreArchivo);
 
-	    try {
-	        Files.copy(rutaOrigen, rutaDestino);
-	    } catch (Exception e) {
-	        System.err.println("Error al copiar el archivo: " + e.getMessage());
+	    if (!Files.exists(rutaDestino)) {
+		    try {
+		        Files.copy(rutaOrigen, rutaDestino);
+		    } catch (Exception e) {
+		        System.err.println("Error al copiar el archivo: " + e.getMessage());
+		    }
 	    }
 	    
 	    return rutaDestino.toString();
